Final Project finished no extra credit.

+++ driver.cpp
@@ -7,9 +7,9 @@ using namespace std;
 
 int main()
 {
-    int exit, num_commit = 0;
+    int exit, num_commit,recent_commit = 0;
     miniGit git;
-    // doublyNode *dNode = new doublyNode;
+    /* doublyNode *dNode = new doublyNode; */
     string exitStr = "", fileName = "";
     ifstream checkFileName;
     bool fileExists = false;
@@ -31,13 +31,21 @@ int main()
             case 1://INIT--DONE
                 //git.init();
                 //dNode = git.addDDnode(num_commit);                          // Initialize empty repository
+                /* dNode->commitNumber = 0;
+                dNode->head = nullptr;
+                dNode->next = nullptr;
+                dNode->previous = nullptr; */
+                //git.init(dNode);
                 git.init(git.addDDnode(num_commit));
                 cout << "New respository initialized." << endl;
                 break;
             
             case 2://ADD--DONE
-                do
-                {
+                if(recent_commit!=num_commit){
+                    cout<<"can't add while not in most recent commit"<<endl;
+                    break;
+                }
+                do{
                     fileExists = false;
                     cout << "Enter file name: " << endl;
                     getline(cin, fileName);
@@ -46,7 +54,11 @@ int main()
                     {
                         cout << "File does not exist." << endl;
                     }
-                    else{break;}
+                    else
+                    {
+                        // fileExists = true;
+                        break;
+                    }
                 }while(checkFileName.fail());
                 if (num_commit > 0 && git.checkFile(fileName, num_commit))
                 {
@@ -60,29 +72,49 @@ int main()
                 break;
             
             case 3://REMOVE--DONE
+            if(recent_commit!=num_commit){
+                    cout<<"can't remove while not in most recent commit"<<endl;
+                    break;
+            }
                 cout << "Enter file name you wish to remove: " << endl;
                 getline(cin, fileName);
                 if (git.checkFile(fileName, num_commit))
                 {
                     git.removeFile(fileName, num_commit);
+                    //git.printDS();
                 }
                 else{cout << "File does not exist within the directory." << endl;}
                 break;
             
             case 4://COMMIT--IN PROGRESS
+            if(recent_commit!=num_commit){
+                    cout<<"can't commit while not in most recent commit"<<endl;
+                    break;
+            }
                 num_commit++;
+                recent_commit++;
                 git.commit(num_commit);
                 break;
             
             case 5://CHECKOUT--IN PROGRESS
-        
-                cout << "Enter the s" << endl;
+                cout << "Are you sure you wish to checkout? *this action will overwrite the current directory* (y/n)"<<endl;
+                char checkout;
+                cin>>checkout;
+                cin.ignore();
+                if(checkout == 'y'){
+                    cout<<"Which commit would you like to checkout?"<<endl;
+                cin>>recent_commit;
+                cin.ignore();
+                git.checkout(recent_commit);
                 break;
-            
+                }
+                else{
+                break;
+                }
             case 6://EXIT--DONE
                 cout << "exit" << endl;
-                git.~miniGit();
                 fs::remove_all(".miniGit");
+                git.~miniGit();
                 break;
             
             case 7:
